partner 페이지가 뜰 수 있도록 만들기
