+++ assets/template/modal-bootstrap5.html
@@ -13,4 +13,5 @@
     {{template "modal-publish" .}}
     {{template "modal-iteminfo" .}}
     {{template "modal-thumbnail" .}}
+    {{template "modal-partner" .}}
 {{end}}
